Fixed behavior with killing processes and changing quantum while running. Kill was removing more then 1 processes due to not having a howmany parameter in the splice function. Also if you changed the quantum to something less than what it was while it was running then it would never context switch because it was only checking for cycles === quantum, changed it to >= to solve this.

+++ scripts/os/kernel.js
@@ -340,7 +340,8 @@ function krnKill(pid) {
   var pcb = null;
   var idx = -1;
 
-  if(_CurrentPCB.pid === pid) {
+  // check to see if it is the current process and make sure it is still running
+  if(_CurrentPCB.pid === pid && _CurrentPCB.finished === false) {
     if(_ReadyQueue.isEmpty())
       _CPU.isExecuting = false;
     _CpuScheduler.killProcess(pid, idx);
@@ -361,8 +362,6 @@ function krnKill(pid) {
   {
     // wasn't found
     _StdOut.putText("Invalid pid.");
-    _StdOut.advanceLine();
-    _StdOut.putText(">");
     return;
   }
